services/motif_service_old/requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.7
imageio==2.32.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lazy_loader==0.3
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.4
onnxruntime==1.18.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
platformdirs==4.0.0
pooch==1.8.1
protobuf==4.25.3
PyMatting==1.1.12
pyreadline3==3.4.1
referencing==0.30.2
rembg==2.0.56
requests==2.31.0
rpds-py==0.12.0
scikit-image==0.22.0
scipy==1.11.3
sympy==1.12
tifffile==2023.9.26
tqdm==4.66.4
urllib3==2.0.7
